Finding outliers in share market data
This assignment concerns the central share market of a small country is facing a problem. An analyst named William is suspected that one specific share trader named Elizabeth is manipulating/hacking the share trading software to make share trades in unusual ways. The only data available to William is about the hourly prices of shares traded. No other information is available to William. The only known indicator of any manipulated distortion (in the share market) is the presence of extreme outliers of share prices that are substantially distant from the share prices of that specific day.

William has arranged a meeting with the presidential body of the share market. In this meeting, William wants to show that something unusual is happening. If he can show anything suspicious, then more investigations will be performed and more data will be provided to him. William is asking for your help in effectively visualizing this data through a sensible graphic. 

Hourly benchmark share prices for one week are provided. Elizabeth did substantial trading on 14th and 18th April only. She was practically inactive on other days. 

I suggested here suitable graphics that helps visualize any notable oddities or suspicious peculiarities.