Transformation of visual information
In a particular research organization, the flow of data is controlled very tightly.  There are three classes of data: Commercial, Technical and Personal. The team of researchers working there have been clearly instructed on what kind of data can be provided to which person. Any person X can request any person Y for any type of data. However, the requested person Y should only provide data to person X if the person Y is indeed permitted to provide that specific class of data to person X. 
I personally went and asked a few associate on what kind of data s/he is permitted to provide to a few other persons. Please note that we do not have data for every single pair of persons. That is intentional.  If the data is not provided for a particular pair of persons, it merely means that data is unavailable. It does not mean that the data transfer is permitted or denied by default. 
Please check how I represent this data visually, and what conclusions I draw using this data.